feat: add ocpi credential and version support
Initial work towards implementing OCPI. Use the /register method
on the manager API to setup OCPI registration. If the manager is
to be the sender pass both the token and url parameters to the
/register endpoint; otherwise pass the token parameter and the
manager will act as the receiver.

Error handling is unlikely to be compliant with the OCPI spec.