feat: improve instrumentation for google cloud mTLS issues
When deployed in google cloud, mTLS is off-loaded to the cloud
load balancer as it needs to terminate TLS. However, the
instrumentation around failures due to this off-load was not
great.

Moved the creation of the initial trace message to middleware.
This allows the middleware that processes the Google Cloud
proxy headers for mTLS to add lots of additional information to
the span.