feat: local CS certificate provider automatically adds cert to store
When deployed in the cloud the cloud load-balancer only provides us
with a hash of the client certificate (along with whether it thinks
it is valid). However, there are other things we need to check in
the certificate so we store the certificate in the database and
look it up based on the certificate hash.

To enable seamless updates to the CSO certificate, the local CS
certificate provider now automatically adds the new certificate to
the store.