feat: added simple admin ui
Allows a charge station to be registered.

Also fixed local charge station certificate provider which would
fail after issuing one certificate as the input stream it was
reading the certificate/key pair from would have been exhausted.

Introduces two new configuration keys:

api.ExternalAddr is the external address on which the CSMS is made
available. This is currently just used for providing connection
information in the admin ui

api.OrgName is the organisation name that is running the CSMS.
This is used as the Organization (O) component of the DN in the
CSRs generated by the admin ui. This should correspond to the
org-name flag passed to the gateway component.