refactor: decouple business logic from transport (mqtt)
This change moves where functionality is provided and adds various
interfaces so that the users of that functionality are decoupled
from the implementation.

The major change is to the code for handling messages being sent
and received from the gateway. The only supported transport for
these messages is still MQTT 5, but this is now very much just an
implementation detail (for the manager - a similar set of changes
would be required in the gateway too).

The core code for working with mqtt has been moved to the
transport/mqtt package which implements interfaces from the
transport package.

The router implementations are now part of the handlers package.
The core router is implemented in the handlers package and there
are functions in the handlers/ocpp16 and handlers/ocpp201
packages for creating a router that is configured for that
version of the OCPP protocol.

The MQTT emitter implementation now owns its own connection to the
MQTT broker, which decouples it from the code for receiving
messages from the broker and means we can easily use the emitter
wherever we need to send messages (whether from a handler or from
an API). This means that the Sender implementation is no longer
required.

The functionality for updating the charge station configuration
is now in its own sync package but is otherwise unchanged.

Basic validation is now performed on the configuration so we
return an error if it is invalid instead of issuing a panic.
Further work is required to format the errors in a way that
makes it easy to understand what configuration is wrong.

The config package now creates more of the types that are used
instead of just returning the raw configuration. This includes
the transport and routers.