refactor: replace transport.Receiver with transport.Listener
New functionality allows multiple connections to be easily established
to the message broker. Connections can either be for all charge
stations, or just for a specific charge station. Sending responses to
the charge station is now the responsibility of the message handler
as some message handlers may not want to send responses to the
charge station.

e.g. An API call may want to wait for the response from the charge
station so it can provide information to the caller. In this
situation the synchronous API call can establish a new connection
to the broker that explicitly listens for messages from the
charge station and waits for the appropriate message to be
received (or a timeout to occur). However, the standard message
processing for that message is still handled by the router.